Edit and Delete
Edit - This works.  I tried to add an image, but could not use the proper file format.
Delete - This works.  I can delete any new legos.  The original legos I added (1-3) get
I tried to add a main menu page but could not redirect the index page to it so just deleted it. I can share via email if you want to try.